+++ lib/extensions.dart
@@ -14,6 +14,7 @@ export 'src/term_similarity/_index.dart'
 export 'src/text_analyzer/_index.dart'
     show
         TokenCollectionExtension,
+        TextAnalysisExtensionsOnJSON,
         KGramExtensionOnTermCollection,
         TextAnalysisExtensionsOnString,
         StringCollectionCollectionExtension,